Add console interaction bits
As the title says, introduce a little helper console command system. So far, it can (besides the builtin help command)
 - exit the server
 - check server and query status
 - force end a query

This also improves the code of the websocket server a tiny bit, to use const references when sending instead of hoping and praying copy elision will carry us where we need to go. I'm the one to blame in this case lol